implement unit test helper in util file for DRY
