+++ tests/kernels/moe/test_modular_oai_triton_moe.py
@@ -36,7 +36,7 @@ from vllm.model_executor.layers.fused_moe.prepare_finalize import (
 from vllm.platforms import current_platform
 from vllm.utils.torch_utils import set_random_seed
 
-from .utils import make_dummy_moe_config
+from .utils import make_dummy_moe_config, shuffle_weight
 
 MNK = [
     (1, 512, 384),
@@ -45,16 +45,6 @@ MNK = [
     (2, 2880, 2880),
     (16, 2880, 2880),
 ]
-
-
-def shuffle_weight(w: torch.Tensor) -> torch.Tensor:
-    """Fold weights to adjacent locations for Triton MoE kernel layout."""
-    shape = w.shape
-    n = shape[-1]
-    first = w[..., : n // 2]
-    second = w[..., n // 2 :]
-    stacked = torch.stack((first, second), dim=-1)
-    return stacked.reshape(shape)
 
 
 def unshuffle_weight(w: torch.Tensor):
